+++ skeleton/src/App/Core/Config.php
@@ -1,0 +1,13 @@
+<?php
+namespace App\Core;
+
+/**
+ *
+ * @author jb
+ *
+ */
+use App\Exception\MinimeException;
+
+class Config extends MinimeConfig
+{
+}