+++ skeleton/src/App/Core/Request.php
@@ -1,0 +1,10 @@
+<?php
+namespace App\Core;
+
+/**
+ *
+ * @author jb
+ */
+Class Request extends MinimeRequest
+{  
+}