+++ skeleton/src/App/Core/Router.php
@@ -1,0 +1,13 @@
+<?php
+namespace App\Core;
+
+use App\Exception\MinimeException;
+
+/**
+ *
+ * @author jb
+ *
+ */
+Class Router extends MinimeRouter
+{
+}