+++ skeleton/src/App/Form/BaseForm.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Form;
+
+/**
+ *
+ * @author jb
+ *
+ */
+class BaseForm extends MinimeForm
+{
+}