+++ skeleton/src/App/Model/ExampleModelInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Model;
+
+interface ExampleModelInterface
+{
+  /**
+   *
+   * @param int $id
+   * @return Ambigous <number, NULL, array>
+   */
+  public function getSomeStuffById($id);
+}