+++ skeleton/src/App/Security/BaseAuth.php
@@ -1,0 +1,11 @@
+<?php
+namespace App\Security;
+
+/**
+ *
+ * @author jb
+ */
+Abstract Class BaseAuth  extends MinimeAuth
+{
+}
+