+++ skeleton/src/App/View/DefaultView.php
@@ -1,0 +1,13 @@
+<?php
+namespace App\View;
+
+require_once __DIR__.'/BaseView.php';
+
+/**
+ *
+ * @author jb
+ */
+Class DefaultView extends BaseView
+{
+  protected $template_name = '/../Templates/Default.tpl.php';
+}