+++ src/App/Core/Container.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * twittee container
+ *
+ * @see http://twittee.org/
+ */
+
+namespace App\Core;
+
+class Container
+{
+  protected $s=array();
+  function __set($k, $c) { $this->s[$k]=$c; }
+  function __get($k) { return $this->s[$k]($this); }
+}